# chore: upgrade Docusaurus, enable Faster