--- a/website/package.json
+++ b/website/package.json
@@ -10,9 +10,10 @@
     "serve": "docusaurus serve"
   },
   "devDependencies": {
-    "@docusaurus/core": "^3.5.2",
-    "@docusaurus/preset-classic": "^3.5.2",
-    "@docusaurus/remark-plugin-npm2yarn": "^3.5.2",
+    "@docusaurus/core": "3.5.2-canary-6121",
+    "@docusaurus/faster": "3.5.2-canary-6121",
+    "@docusaurus/preset-classic": "3.5.2-canary-6121",
+    "@docusaurus/remark-plugin-npm2yarn": "3.5.2-canary-6121",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-markdown": "^9.0.0",