Docusaurus v3.6 had [Docusaurus Faster](https://github.com/facebook/docusaurus/issues/10556) options.

I'm testing it on a few sites before the release.

## Benchmark

Cold builds - much faster:

```
hyperfine --prepare 'yarn --cwd website clear' --runs 3 'DOCUSAURUS_FASTER=false yarn build:docusaurus' 'DOCUSAURUS_FASTER=true yarn build:docusaurus'


Benchmark 1: DOCUSAURUS_FASTER=false yarn build:docusaurus
  Time (mean ± σ):     26.333 s ±  0.075 s    [User: 67.591 s, System: 7.877 s]
  Range (min … max):   26.248 s … 26.388 s    3 runs

Benchmark 2: DOCUSAURUS_FASTER=true yarn build:docusaurus
  Time (mean ± σ):      8.062 s ±  0.348 s    [User: 15.819 s, System: 1.906 s]
  Range (min … max):    7.660 s …  8.272 s    3 runs

Summary
  DOCUSAURUS_FASTER=true yarn build:docusaurus ran
    3.27 ± 0.14 times faster than DOCUSAURUS_FASTER=false yarn build:docusaurus
```

Warm rebuilds - same time (for now):

```bash
DOCUSAURUS_FASTER=false yarn build
hyperfine --runs 3 'DOCUSAURUS_FASTER=false yarn build:docusaurus' 'DOCUSAURUS_FASTER=true yarn build:docusaurus'


Benchmark 1: DOCUSAURUS_FASTER=false yarn build:docusaurus
  Time (mean ± σ):      7.934 s ±  0.356 s    [User: 10.059 s, System: 1.205 s]
  Range (min … max):    7.727 s …  8.345 s    3 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 2: DOCUSAURUS_FASTER=true yarn build:docusaurus
  Time (mean ± σ):      8.119 s ±  0.549 s    [User: 15.756 s, System: 1.909 s]
  Range (min … max):    7.656 s …  8.725 s    3 runs

Summary
  DOCUSAURUS_FASTER=false yarn build:docusaurus ran
    1.02 ± 0.08 times faster than DOCUSAURUS_FASTER=true yarn build:docusaurus
```