@slorber 2024-10-25
Just realized I'm upgrading the Babel website to use SWC 😅 hope you are fine with this